+++ code/spikes/deno.000/src.api.deno/mod.ts
@@ -1,3 +1,4 @@
 import { route as subhosting } from './route.subhosting/mod.ts';
 
 export const Routes = { subhosting } as const;
+export default Routes;
