+++ code/spikes/deno.000/src.api.openai/mod.ts
@@ -4,7 +4,7 @@ import { DEFAULTS, Is, OpenAI, type t } from './u.common.ts';
 /**
  * OpenAI route setup.
  */
-export default function init(path: string, app: t.HonoApp, env: { apiKey: string }) {
+export default function routes(path: string, app: t.HonoApp, env: { apiKey: string }) {
   const { apiKey } = env;
   const openai = new OpenAI({ apiKey });
