+++ code/spikes/deno.000/src.api.openai/u.common.ts
@@ -1,7 +1,8 @@
 import OpenAI from 'npm:openai';
 export { OpenAI };
+export type * as t from './types.ts';
 
-export * from '../common/mod.ts';
+export * from '../src/common.ts';
 export * from './u.is.ts';
 
 export const DEFAULTS = { model: 'gpt-3.5-turbo' } as const;
