+++ code/spikes/deno.000/src/common.libs.ts
