+++ code/spikes/deno.000/src/common.ts
@@ -1,7 +1,7 @@
-export type * as t from '../types.ts';
-export * from './libs.ts';
+export type * as t from './types.ts';
+export * from './common.libs.ts';
 
-export { EnvVars } from '../env.ts';
+export { EnvVars } from './env.ts';
 
 /**
  * Helpers
