+++ code/spikes/deno.000/src/types.ts
@@ -1,2 +1,1 @@
 export type { HonoApp } from '../src.api.deno/types.ts';
-export type * from './r.openai/t.ts';
